add LogEntryVariant to helpers
