+++ src/pages/namespace/Instances/Detail/Main/Logs/Entry.tsx
@@ -1,28 +1,11 @@
-import { ComponentProps, ComponentPropsWithoutRef, forwardRef } from "react";
+import { ComponentPropsWithoutRef, forwardRef } from "react";
+import { formatLogTime, logLevelToLogEntryVariant } from "~/util/helpers";
 
 import { LogEntry } from "~/design/Logs";
 import { LogEntryType } from "~/api/logs/schema";
-import { formatLogTime } from "~/util/helpers";
 import { useLogsPreferencesVerboseLogs } from "~/util/store/logs";
 
-type LogEntryVariant = ComponentProps<typeof LogEntry>["variant"];
-type logLevel = LogEntryType["level"];
-
-const logLevelToLogEntryVariant = (level: logLevel): LogEntryVariant => {
-  switch (level) {
-    case "error":
-      return "error";
-    case "info":
-      return "info";
-    case "debug":
-      return undefined;
-    default:
-      break;
-  }
-};
-
 type LogEntryProps = ComponentPropsWithoutRef<typeof LogEntry>;
-
 type Props = { logEntry: LogEntryType } & LogEntryProps;
 
 export const Entry = forwardRef<HTMLDivElement, Props>(
