+++ src/pages/namespace/Monitoring/Logs/Entry.tsx
@@ -1,28 +1,10 @@
-import { ComponentProps, ComponentPropsWithoutRef, forwardRef } from "react";
+import { ComponentPropsWithoutRef, forwardRef } from "react";
+import { formatLogTime, logLevelToLogEntryVariant } from "~/util/helpers";
 
 import { LogEntry } from "~/design/Logs";
 import { NamespaceLogSchemaType } from "~/api/namespaces/schema";
-import { formatLogTime } from "~/util/helpers";
-
-type LogEntryVariant = ComponentProps<typeof LogEntry>["variant"];
-type logLevel = NamespaceLogSchemaType["level"];
-
-// mage log level a more generic type, also this method
-const logLevelToLogEntryVariant = (level: logLevel): LogEntryVariant => {
-  switch (level) {
-    case "error":
-      return "error";
-    case "info":
-      return "info";
-    case "debug":
-      return undefined;
-    default:
-      break;
-  }
-};
 
 type LogEntryProps = ComponentPropsWithoutRef<typeof LogEntry>;
-
 type Props = { logEntry: NamespaceLogSchemaType } & LogEntryProps;
 
 export const Entry = forwardRef<HTMLDivElement, Props>(
